add support for fusemap vendor overrides
