+++ otp/otp_linux.go
@@ -6,6 +6,7 @@
 // Use of this source code is governed by the license
 // that can be found in the LICENSE file.
 //
+//go:build linux
 // +build linux
 
 package otp
